moved oyster store manipulation logic to the store file
