+++ src/lib/page-components/oyster/inventory/modals/StopJobModal.svelte
@@ -46,7 +46,7 @@
 	};
 
 	const handleOnTimerEnd = async () => {
-		await handleJobStatusOnStopTimerEnd(jobData);
+		handleJobStatusOnStopTimerEnd(jobData);
 	};
 
 	$: modalTitle =
