+++ src/lib/utils/services/oysterServices.ts
@@ -1,8 +1,16 @@
 import type { BigNumber, Bytes } from 'ethers';
 import {
-	OYSTER_RATE_METADATA,
-	OYSTER_RATE_SCALING_FACTOR
-} from '$lib/utils/constants/oysterConstants';
+	addFundsToJobInOysterStore,
+	cancelRateReviseInOysterStore,
+	createNewJobInOysterStore,
+	initiateRateReviseInOysterStore,
+	stopJobInOysterStore,
+	updateAmountToBeSettledForJobInOysterStore,
+	updateApprovedFundsInOysterStore,
+	updateJobRateInOysterStore,
+	updateJobStatusOnTimerEndInOysterStore,
+	withdrawFundsFromJobInOysterStore
+} from '$lib/data-stores/oysterStore';
 import {
 	addFundsToOysterJob,
 	approveFundsForOysterJobAdd,
@@ -17,28 +25,11 @@ import {
 
 import { BIG_NUMBER_ZERO } from '$lib/utils/constants/constants';
 import type { OysterInventoryDataModel } from '$lib/types/oysterComponentType';
-import type { OysterStore } from '$lib/types/storeTypes';
-import { oysterStore } from '$lib/data-stores/oysterStore';
-import { parseMetadata } from '$lib/utils/data-modifiers/oysterModifiers';
 
 export async function handleClaimAmountFromOysterJob(jobId: Bytes) {
 	try {
 		await settleOysterJob(jobId);
-		oysterStore.update((value: OysterStore) => {
-			return {
-				...value,
-				merchantJobsData: value.merchantJobsData.map((job) => {
-					if (job.id === jobId) {
-						console.log('job :>> ', job);
-						return {
-							...job,
-							amountToBeSettled: BIG_NUMBER_ZERO
-						};
-					}
-					return job;
-				})
-			};
-		});
+		updateAmountToBeSettledForJobInOysterStore(jobId, BIG_NUMBER_ZERO);
 	} catch (e) {
 		console.log('e :>> ', e);
 	}
@@ -47,12 +38,7 @@ export async function handleClaimAmountFromOysterJob(jobId: Bytes) {
 export async function handleApproveFundForOysterJob(amount: BigNumber) {
 	try {
 		await approveFundsForOysterJobAdd(amount);
-		oysterStore.update((value) => {
-			return {
-				...value,
-				allowance: amount
-			};
-		});
+		updateApprovedFundsInOysterStore(amount);
 	} catch (e) {
 		console.log('e :>> ', e);
 	}
@@ -66,38 +52,7 @@ export async function handleFundsAddToJob(
 	const { id } = jobData;
 	try {
 		const txn = await addFundsToOysterJob(id, amount);
-		const nowTime = Date.now() / 1000;
-		const modifiedJobData = {
-			...jobData,
-			totalDeposit: jobData.totalDeposit.add(amount),
-			balance: jobData.balance.add(amount),
-			durationLeft: jobData.durationLeft + duration,
-			endEpochTime: jobData.endEpochTime + duration,
-			depositHistory: [
-				{
-					amount,
-					id: txn.id,
-					txHash: txn.hash,
-					timestamp: nowTime,
-					isWithdrawal: false,
-					transactionStatus: 'deposit'
-				},
-				...jobData.depositHistory
-			]
-		};
-		oysterStore.update((value: OysterStore) => {
-			return {
-				...value,
-				allowance: value.allowance.sub(amount),
-				jobsData: value.jobsData.map((job) => {
-					if (job.id === id) {
-						return modifiedJobData;
-					}
-					return job;
-				})
-			};
-		});
-		// return modifiedJobData;
+		addFundsToJobInOysterStore(id, txn, jobData, amount, duration);
 	} catch (e) {
 		console.log('e :>> ', e);
 		return jobData;
@@ -111,34 +66,7 @@ export async function handleFundsWithdrawFromJob(
 	const { id } = jobData;
 	try {
 		const txn = await withdrawFundsFromOysterJob(id, amount);
-		const nowTime = Date.now() / 1000;
-		const modifiedJobData = {
-			...jobData,
-			totalDeposit: jobData.totalDeposit.sub(amount),
-			balance: jobData.balance.sub(amount),
-			depositHistory: [
-				{
-					amount,
-					id: txn.id,
-					txHash: txn.hash,
-					timestamp: nowTime,
-					isWithdrawal: true,
-					transactionStatus: 'withdrawal'
-				},
-				...jobData.depositHistory
-			]
-		};
-		oysterStore.update((value: OysterStore) => {
-			return {
-				...value,
-				jobsData: value.jobsData.map((job) => {
-					if (job.id === id) {
-						return modifiedJobData;
-					}
-					return job;
-				})
-			};
-		});
+		withdrawFundsFromJobInOysterStore(id, txn, jobData, amount);
 	} catch (e) {
 		console.log('e :>> ', e);
 		return jobData;
@@ -149,31 +77,10 @@ export async function handleInitiateRateRevise(
 	jobData: OysterInventoryDataModel,
 	newRate: BigNumber
 ) {
-	const { rateReviseWaitingTime } = OYSTER_RATE_METADATA;
 	const { id } = jobData;
 	try {
 		await initiateRateReviseOysterJob(id, newRate);
-		const nowTime = Date.now() / 1000;
-		const modifiedJobData = {
-			...jobData,
-			reviseRate: {
-				newRate: newRate,
-				rateStatus: 'pending',
-				stopStatus: newRate.gt(BIG_NUMBER_ZERO) ? 'disabled' : 'pending',
-				updatesAt: nowTime + rateReviseWaitingTime
-			}
-		};
-		oysterStore.update((value: OysterStore) => {
-			return {
-				...value,
-				jobsData: value.jobsData.map((job) => {
-					if (job.id === id) {
-						return modifiedJobData;
-					}
-					return job;
-				})
-			};
-		});
+		initiateRateReviseInOysterStore(id, jobData, newRate);
 	} catch (e) {
 		console.log('e :>> ', e);
 	}
@@ -183,50 +90,15 @@ export async function handleCancelRateRevise(jobData: OysterInventoryDataModel) 
 	const { id } = jobData;
 	try {
 		await cancelRateReviseOysterJob(id);
-		const modifiedJobData = {
-			...jobData,
-			reviseRate: undefined
-		};
-		oysterStore.update((value: OysterStore) => {
-			return {
-				...value,
-				jobsData: value.jobsData.map((job) => {
-					if (job.id === id) {
-						return modifiedJobData;
-					}
-					return job;
-				})
-			};
-		});
+		cancelRateReviseInOysterStore(id, jobData);
 	} catch (e) {
 		console.log('e :>> ', e);
 	}
 }
 
-export async function handleJobStatusOnStopTimerEnd(jobData: OysterInventoryDataModel) {
-	const { id } = jobData;
+export function handleJobStatusOnStopTimerEnd(jobData: OysterInventoryDataModel) {
 	try {
-		const nowTime = Date.now() / 1000;
-		const modifiedJobData = {
-			...jobData,
-			reviseRate: {
-				newRate: BIG_NUMBER_ZERO,
-				rateStatus: 'pending',
-				stopStatus: 'completed',
-				updatesAt: nowTime
-			}
-		};
-		oysterStore.update((value: OysterStore) => {
-			return {
-				...value,
-				jobsData: value.jobsData.map((job) => {
-					if (job.id === id) {
-						return modifiedJobData;
-					}
-					return job;
-				})
-			};
-		});
+		updateJobStatusOnTimerEndInOysterStore(jobData);
 	} catch (e) {
 		console.log('e :>> ', e);
 	}
@@ -239,21 +111,7 @@ export async function handleFinaliseRateRevise(
 	const { id } = jobData;
 	try {
 		await finaliseRateReviseOysterJob(id);
-		oysterStore.update((value) => {
-			return {
-				...value,
-				jobsData: value.jobsData.map((job) => {
-					if (job.id === id) {
-						return {
-							...job,
-							rate: newRate,
-							reviseRate: undefined
-						};
-					}
-					return job;
-				})
-			};
-		});
+		updateJobRateInOysterStore(id, newRate);
 	} catch (e) {
 		console.log('e :>> ', e);
 	}
@@ -262,40 +120,8 @@ export async function handleFinaliseRateRevise(
 export async function handleConfirmJobStop(jobData: OysterInventoryDataModel) {
 	const { id } = jobData;
 	try {
-		const tx = await stopOysterJob(id);
-		const nowTime = Date.now() / 1000;
-		const modifiedJobData = {
-			...jobData,
-			live: false,
-			refund: jobData.balance,
-			balance: BIG_NUMBER_ZERO,
-			status: 'stopped',
-			rate: BIG_NUMBER_ZERO,
-			reviseRate: undefined,
-			endEpochTime: nowTime,
-			depositHistory: [
-				{
-					amount: jobData.balance,
-					id: tx.id,
-					txHash: tx.hash,
-					timestamp: nowTime,
-					isWithdrawal: true,
-					transactionStatus: 'stopped'
-				},
-				...jobData.depositHistory
-			]
-		};
-		oysterStore.update((value) => {
-			return {
-				...value,
-				jobsData: value.jobsData.map((job) => {
-					if (job.id === id) {
-						return modifiedJobData;
-					}
-					return job;
-				})
-			};
-		});
+		const txn = await stopOysterJob(id);
+		stopJobInOysterStore(id, txn, jobData);
 	} catch (e) {
 		console.log('e :>> ', e);
 		return jobData;
@@ -317,60 +143,17 @@ export async function handleCreateJob(
 			rate,
 			balance
 		);
-
-		const txHash = txn.hash;
-		const jobOpenEvent = approveReciept.events?.find((event: any) => event.event === 'JobOpened');
-		const jobId = jobOpenEvent?.args?.job;
-
-		const nowTime = Date.now() / 1000;
-
-		const { enclaveUrl, instance, region, vcpu, memory } = parseMetadata(metadata);
-		const newJob: OysterInventoryDataModel = {
-			id: jobId,
-			provider: {
-				name: provider?.name || '',
-				address: provider.address
-			},
+		createNewJobInOysterStore(
+			txn,
+			approveReciept,
 			owner,
 			metadata,
-			enclaveUrl,
-			instance,
-			region,
-			vcpu,
-			memory,
-			amountUsed: BIG_NUMBER_ZERO,
-			refund: BIG_NUMBER_ZERO,
+			provider,
 			rate,
-			downScaledRate: rate.div(OYSTER_RATE_SCALING_FACTOR),
 			balance,
-			totalDeposit: balance,
-			live: true,
-			lastSettled: nowTime,
-			createdAt: nowTime,
-			endEpochTime: nowTime + durationInSec,
-			durationLeft: durationInSec,
-			durationRun: 0,
-			status: 'running',
-			depositHistory: [
-				{
-					amount: balance,
-					id: txHash,
-					txHash: txHash,
-					timestamp: nowTime,
-					isWithdrawal: false,
-					transactionStatus: 'deposit'
-				}
-			],
-			amountToBeSettled: BIG_NUMBER_ZERO,
-			settlementHistory: []
-		};
-		oysterStore.update((value) => {
-			return {
-				...value,
-				jobsData: [newJob, ...value.jobsData],
-				allowance: value.allowance.sub(balance)
-			};
-		});
+			durationInSec
+		);
+
 		return true;
 	} catch (e) {
 		console.log('e :>> ', e);
