+++ src/routes/oyster/operator/jobs/+page.svelte
@@ -3,7 +3,7 @@
 	import { getJobStatuses } from '$lib/controllers/httpController';
 	import { getOysterMerchantJobsFromSubgraph } from '$lib/controllers/subgraphController';
 	import { chainStore } from '$lib/data-stores/chainProviderStore';
-	import { oysterStore } from '$lib/data-stores/oysterStore';
+	import { updateMerchantJobsInOysterStore } from '$lib/data-stores/oysterStore';
 	import { connected, walletStore } from '$lib/data-stores/walletProviderStore';
 	import OysterMerchantJobs from '$lib/page-components/oyster/operator/OysterOperatorJobs.svelte';
 	import { modifyOysterJobData } from '$lib/utils/data-modifiers/oysterModifiers';
@@ -28,21 +28,9 @@
 		const merchantJobs = await modifyOysterJobData(merchantJobsFromSubgraph);
 		// updating the oyster store with merchant jobs
 		if (merchantJobs !== null) {
-			oysterStore.update((value) => {
-				return {
-					...value,
-					merchantJobsData: merchantJobs,
-					merchantJobsLoaded: true
-				};
-			});
+			updateMerchantJobsInOysterStore(merchantJobs);
 		} else {
-			oysterStore.update((value) => {
-				return {
-					...value,
-					merchantJobsData: [],
-					merchantJobsLoaded: true
-				};
-			});
+			updateMerchantJobsInOysterStore([]);
 		}
 	}
